Don't create the database during the setup.  Build it during make.
The database tables and the database were being setup during the
calibTables cmt setup, but this had the problem of trying to write to an
area that might not be owned by the user.  The requirements have been
changed so that the database is now setup as part of the build procedure
(by the make command).