Hand rearrange the first half of the collection plane.
The first half of the X plane now looks pretty good. There are still some
channels out of place, and the even/odd problem persists, but the major
blocks are in pretty good shape.  One on-going problem is that it seems
that there are probably still some channels that are attached to wires, but
which don't have a wire assigned.