Reverse the wire order on the X plane.
The initial X plane mapping was done based on proton tracks in the neutron
beam, but involved a misunderstanding of the relation between the beam and
detector geometry.  Thee preliminary mapping has been used to compare track
geometries on the U, V, and X planes.  It is apparent that the wire
ordering on the X plane has been reversed.  The U and V plane wire
orderings were carefully checked against drawings, cosmic tracks, and laser
tracks during the February engineering run (when the X plane was not
working).