A validated wire mapping.
The U, V and X wire mappings have been cross validated, and since the U and
V mappings were validated using laser data, this should be pretty close to
a final wire map.  The large blocks of wires (e.g. groups of 16 wires) in
the X mapping may still have be shifted by a few wires relative to their
correct position, but the wires are in the right order.